--- a/man/lang.Rd
+++ b/man/lang.Rd
@@ -14,7 +14,7 @@ new_language(head, tail = NULL)
 symbol, call, or a function.}
 
 \item{...}{Arguments to the call either in or out of a list. These dots
-support \link[=tidy-dots]{tidy dots} features.}
+support \link[=tidy-dots]{tidy dots} features. Empty arguments are preserved.}
 
 \item{.ns}{Namespace with which to prefix \code{.fn}. Must be a string
 or symbol.}