--- a/src/Makevars
+++ b/src/Makevars
@@ -36,6 +36,7 @@ internal-files = \
         internal/eval-tidy.c \
         internal/expr-interp.c \
         internal/expr-interp-rotate.c \
+        internal/fn.c \
         internal/internal.c \
         internal/quo.c \
         internal/utils.c