--- a/src/internal/call.c
+++ b/src/internal/call.c
@@ -32,6 +32,9 @@ static sexp* rlang_call2(sexp* fn, sexp* args, sexp* ns) {
     if (!r_is_string(ns, NULL)) {
       r_abort("`ns` must be a string");
     }
+    if (r_typeof(fn) != r_type_symbol) {
+      r_abort("`fn` must be a string or symbol when a namespace is supplied");
+    }
     ns = r_sym(r_chr_get_c_string(ns, 0));
     fn = KEEP_N(r_call3(r_namespace_sym, ns, fn), n_protect);
   }