--- a/src/internal/internal.c
+++ b/src/internal/internal.c
@@ -7,6 +7,9 @@ sexp* rlang_zap = NULL;
 sexp* as_list_call = NULL;
 sexp* as_list_s4_call = NULL;
 
+sexp* rlang_objs_keep = NULL;
+sexp* rlang_objs_trailing = NULL;
+
 void rlang_init_dots();
 void rlang_init_expr_interp();
 void rlang_init_eval_tidy();
@@ -23,6 +26,13 @@ void rlang_init_internal() {
 
   as_list_s4_call = r_parse("as(x, 'list')");
   r_mark_precious(as_list_s4_call);
+
+
+  rlang_objs_keep = r_chr("keep");
+  r_mark_precious(rlang_objs_keep);
+
+  rlang_objs_trailing = r_chr("trailing");
+  r_mark_precious(rlang_objs_trailing);
 
   /* dots.c - enum dots_expansion_op */
   RLANG_ASSERT(OP_DOTS_MAX == DOTS_CAPTURE_TYPE_MAX * EXPANSION_OP_MAX);