… NAMESPACE
… man/pairlist2.Rd