* New `pairlist2()` function to create pairlists with splicing support. It's mainly meant to support `call2()` but may also be useful for `new_function()`. It preserves empty arguments since those are useful in calls and formals lists.

* `call2()` is reimplemented in C and uses `pairlist2()` for capturing dots. This brings its performance closer to that of `base::call()`:

  ```r
  bench::mark(
    base = call("fn"),
    before = call2_before("fn"),
    after = call2("fn")
  )
  #>   expression     min    mean  median    max `itr/sec` mem_alloc  n_gc n_itr
  #>   <bch:expr> <bch:t> <bch:t> <bch:t> <bch:>     <dbl> <bch:byt> <dbl> <int>
  #> 1 base          48ns  76.8ns    59ns 14.7µs 13013347.        0B     0 10000
  #> 2 before      11.2µs  14.5µs  13.4µs 94.3µs    68865.        0B     3  9997
  #> 3 after        698ns 867.4ns   806ns 24.9µs  1152811.        0B     0 10000
  
  args <- list(foo = missing_arg(), bar = 2)
  bench::mark(
    base = as.call(c(list(as.name("fn")), args)),
    before = call2_before("fn", !!!args),
    after = call2("fn", !!!args)
  )
  #>   expression     min    mean  median     max `itr/sec` mem_alloc  n_gc n_itr
  #>   <bch:expr> <bch:t> <bch:t> <bch:t> <bch:t>     <dbl> <bch:byt> <dbl> <int>
  #> 1 base         777ns  1.02µs   892ns  15.2µs   982750.        0B     0 10000
  #> 2 before        13µs 16.76µs  15.7µs 104.7µs    59670.        0B     2  9998
  #> 3 after        1.4µs  1.74µs   1.6µs  28.5µs   573699.        0B     1  9999
  ```

  Like `pairlist2()`, it preserves empty arguments:

  ```
  call2("[", quote(x), , j = 2)
  #> x[, j = 2]
  ```

* `new_function()` is now implemented in C as well.